adapt 7.5
